moved the dummy test case to its own test file
